move AddStaticImport quickfix to QuickFixProcessor
